Copy+Delete because filesystems
Turns out that ‘fs::rename()’ calls ‘rename(2)’ under the hood.  This is
a problem because the temporary directory we use is typically located on
a different filesystem than your home directory for example, and so this
would fail.  The shitty workaround here is to copy the files and then
delete them.

But it just keeps getting better!  Turns out that despite the fact that
‘everything is a file’ on *NIX, we cannot just call ‘fs::copy()’ or
‘fs::remove_file()’ on directories; that would be too smart.