Add MMV_NAME and MCP_NAME compile-time variables
Due to the sheer popularity of the "mmv" name it might conflict with
other programs installed on the user's computer.

As such the MMV_NAME and MCP_NAME environmental variables can control
the program's name at compile time, allowing the user to avoid name
conflicts with other programs.

If the variables are not set then they will silently default back to
"mmv" and "mcp" respectively.

Signed-off-by: Luca Matei Pintilie <luca@lucamatei.com>
Co-authored-by: Thomas Voss <mail@thomasvoss.com>